+++ CHANGELOG.md
@@ -98,6 +98,7 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 ## 1.6.0 - 2020-06-30
 
 ### Added
+
 - Make `ExposureType.OUTSIDE_TRUSTED_DOMAINS` constant available.
 - `sdk.securitydata` methods
     - `stream_file_by_sha256()`